add remove cache on multiple pages
fix imports
